+++ lib/view/constants/calendar_constants.dart
@@ -1,0 +1,4 @@
+class CalendarConstants {
+  static DateTime get calendarFirstDay => DateTime.utc(1900, 1, 1);
+  static DateTime get calendarEndDay => DateTime.utc(2100, 12, 31);
+}
